refactor(tasks): split tasks.rs into submodules (#541)
The 2008-line `src/engine/tasks.rs` had grown to cover seven distinct
concerns: the `define_encode_task!` macro, the atomic file-write helper,
`TaskContext` and decode plumbing, the two-phase encode pipeline, buffer
encode tasks, file-output encode tasks, and batch processing. This commit
breaks it apart into a `tasks/` directory module:

  - macros.rs     — `define_encode_task!` (kept internal via `#[macro_use]`)
  - atomic.rs     — `write_encoded_output` atomic temp-file + rename helper
  - context.rs    — `TaskContext`, `BatchResult`, `decode_internal_from_parts`,
                    format helpers
  - processing.rs — `EncodeContext`, `PreparedEncode`,
                    `prepare_for_encode` / `encode_prepared` /
                    `process_and_encode_from_parts`
  - encode.rs     — `EncodeTask`, `EncodeWithMetricsTask`,
                    `EncodeTargetBytesTask` (binary search) +
                    `TargetBytesOutput`
  - write.rs      — `WriteFileTask`, `WriteFileWithMetricsTask`,
                    `UnifiedEncodeTask`, `UnifiedEncodeToFileTask`
  - batch.rs      — `BatchFileContext`, `process_batch_file`,
                    `BatchTask`, `BatchWithMetricsTask`
  - tests.rs      — non-NAPI unit + regression tests

Internal seams (the encode pipeline structs, decode helpers, atomic
helper) are `pub(super)` so the directory's externally visible surface is
exactly the set of task structs (and `BatchResult` under `cfg(napi)`)
re-exported from `mod.rs`. `api.rs` and `engine.rs` continue to import
from `crate::engine::tasks::…` unchanged.

No behavior change. Verified with `cargo fmt --check`,
`cargo clippy --no-default-features -- -D warnings`,
`cargo test --no-default-features` (276 lib tests + 8 property tests),
`npm run build`, and `npm test`.

Closes #541
