--- src/engine/tasks.rs
@@ -1,2008 +1,0 @@
-#![cfg_attr(not(feature = "napi"), allow(dead_code))]
-
-// src/engine/tasks.rs
-//
-// Async task implementations for NAPI.
-// These tasks run in background threads and don't block Node.js main thread.
-//
-// Common fields are consolidated into `TaskContext`. The `define_encode_task!`
-// macro generates task structs with their NAPI `Task` trait impls, eliminating
-// the per-struct boilerplate that was previously repeated 7+ times.
-
-use super::api::MetadataPolicy;
-use super::firewall::FirewallConfig;
-use crate::engine::decoder::{
-    check_dimensions, decode_image, detect_format, ensure_dimensions_safe,
-};
-use crate::engine::encoder::{
-    embed_exif_jpeg, embed_exif_png, embed_exif_webp, encode_avif, encode_jpeg_with_settings,
-    encode_png, encode_webp,
-};
-#[allow(unused_imports)]
-use crate::engine::io::{exif_has_gps, extract_exif_raw, extract_icc_profile, has_exif, Source};
-use crate::engine::memory;
-use crate::engine::pipeline::{apply_ops_tracked, ColorState, IccState};
-#[cfg(feature = "napi")]
-use crate::engine::pool;
-#[allow(unused_imports)]
-use crate::error::{ErrorCategory, LazyImageError};
-use crate::ops::{Operation, OutputFormat};
-use crate::PROCESSING_METRICS_VERSION;
-use image::{DynamicImage, GenericImageView, ImageFormat};
-#[cfg(feature = "napi")]
-use napi::bindgen_prelude::*;
-#[cfg(feature = "napi")]
-use napi::{bindgen_prelude::BufferSlice, Env, Task};
-#[cfg(feature = "napi")]
-use rayon::prelude::*;
-use std::borrow::Cow;
-use std::sync::Arc;
-use std::time::Instant;
-
-// Resource usage delegates to the centralized platform module.
-use super::platform;
-type ResourceUsage = platform::ResourceUsage;
-
-fn get_resource_usage() -> Option<ResourceUsage> {
-    platform::get_resource_usage()
-}
-
-fn detect_input_format(bytes: &[u8]) -> Option<String> {
-    detect_format(bytes).map(format_to_string)
-}
-
-fn format_to_string(fmt: ImageFormat) -> String {
-    match fmt {
-        ImageFormat::Jpeg => "jpeg",
-        ImageFormat::Png => "png",
-        ImageFormat::WebP => "webp",
-        ImageFormat::Avif => "avif",
-        ImageFormat::Gif => "gif",
-        ImageFormat::Bmp => "bmp",
-        ImageFormat::Ico => "ico",
-        ImageFormat::Tiff => "tiff",
-        other => other.to_mime_type(),
-    }
-    .to_string()
-}
-
-// Re-export BatchResult for api.rs
-#[cfg(feature = "napi")]
-#[napi(object)]
-pub struct BatchResult {
-    pub source: String,
-    pub success: bool,
-    pub error: Option<String>,
-    pub output_path: Option<String>,
-    pub error_code: Option<String>,
-    pub error_category: Option<ErrorCategory>,
-    pub effective_concurrency: Option<u32>,
-    pub auto_concurrency: Option<bool>,
-}
-
-// ============================================================================================
-// TaskContext — common fields shared by all single-image encode tasks
-// ============================================================================================
-
-/// Common fields shared by all single-image encode/write tasks.
-///
-/// Consolidates source, decoded image, operations, format, ICC/EXIF data,
-/// auto-orient flag, metadata policy, and firewall config into a single struct.
-/// Each task struct embeds a `TaskContext` instead of repeating these fields.
-pub struct TaskContext {
-    pub source: Option<Source>,
-    /// Decoded image wrapped in Arc. decode() returns Cow::Borrowed pointing here,
-    /// enabling true Copy-on-Write in apply_ops (no deep copy for format-only conversion).
-    pub decoded: Option<Arc<DynamicImage>>,
-    pub ops: Vec<Operation>,
-    pub format: OutputFormat,
-    pub icc_profile: Option<Arc<Vec<u8>>>,
-    /// Whether the input originally had an ICC profile (even if stripped)
-    pub icc_present: bool,
-    /// Raw EXIF data extracted from source image (for preservation)
-    pub exif_data: Option<Arc<Vec<u8>>>,
-    pub auto_orient: bool,
-    /// Single source of truth for metadata preservation decisions
-    pub metadata_policy: MetadataPolicy,
-    pub firewall: FirewallConfig,
-    /// Last error that occurred during compute (for use in reject)
-    #[cfg(feature = "napi")]
-    pub(crate) last_error: Option<LazyImageError>,
-}
-
-fn decode_internal_from_parts<'a>(
-    source: Option<&'a Source>,
-    decoded: Option<&'a Arc<DynamicImage>>,
-    firewall: &FirewallConfig,
-) -> std::result::Result<Cow<'a, DynamicImage>, LazyImageError> {
-    // Prefer already decoded image (already validated)
-    // Return borrowed reference - no deep copy until mutation is needed
-    if let Some(img_arc) = decoded {
-        // Always-on base safety limits (even without sanitize())
-        firewall.enforce_base_limits(img_arc.width(), img_arc.height())?;
-        check_dimensions(img_arc.width(), img_arc.height())?;
-        firewall.enforce_pixels(img_arc.width(), img_arc.height())?;
-        return Ok(Cow::Borrowed(img_arc.as_ref()));
-    }
-
-    // Get bytes from source - zero-copy for Memory and Mapped sources
-    let bytes = match source {
-        Some(source) => {
-            if let Some(bytes) = source.as_bytes() {
-                bytes
-            } else {
-                // Path sources require loading first - this should not happen in normal flow
-                // as from_path() converts Path to Mapped. If this occurs, it's a programming error.
-                return Err(LazyImageError::decode_failed(
-                    "Path source requires loading first. Use Mapped source (from_path) instead."
-                        .to_string(),
-                ));
-            }
-        }
-        None => {
-            return Err(LazyImageError::source_consumed());
-        }
-    };
-
-    // Always-on base metadata size check (even without sanitize())
-    firewall.scan_metadata_base(bytes)?;
-
-    firewall.enforce_source_len(bytes.len())?;
-    firewall.scan_metadata(bytes)?;
-
-    ensure_dimensions_safe(bytes)?;
-
-    let (img, _detected_format) = decode_image(bytes)?;
-
-    // Security check: reject decompression bombs
-    let (w, h) = img.dimensions();
-    // Always-on base safety limits (even without sanitize())
-    firewall.enforce_base_limits(w, h)?;
-    check_dimensions(w, h)?;
-    firewall.enforce_pixels(w, h)?;
-
-    Ok(Cow::Owned(img))
-}
-
-/// Borrowed parameter bundle for the process + encode pipeline.
-///
-/// Replaces a 10-argument signature on `process_and_encode_from_parts` and an
-/// 8-argument signature on `encode_prepared` with a single struct. Every field
-/// is `Copy`, so the struct itself is `Copy` — that lets callers reuse a base
-/// context and vary only one field with struct-update syntax (used by the
-/// byte-budget binary search, which keeps every field constant except the
-/// per-iteration `OutputFormat`).
-///
-/// `encode_prepared` ignores the prepare-only fields (`source`, `decoded`,
-/// `ops`, `icc_present`); they live here rather than in a separate struct so
-/// every callsite builds exactly one context object.
-#[derive(Clone, Copy)]
-struct EncodeContext<'a> {
-    source: Option<&'a Source>,
-    decoded: Option<&'a Arc<DynamicImage>>,
-    ops: &'a [Operation],
-    format: &'a OutputFormat,
-    icc_profile: Option<&'a Arc<Vec<u8>>>,
-    icc_present: bool,
-    exif_data: Option<&'a Arc<Vec<u8>>>,
-    auto_orient: bool,
-    policy: &'a MetadataPolicy,
-    firewall: &'a FirewallConfig,
-}
-
-/// Cached state from decode + ops, reusable across multiple encode passes.
-///
-/// Used by [`EncodeTargetBytesTask`] to avoid re-decoding and re-applying ops
-/// for every quality candidate in the binary search. The memory permit is held
-/// across all encode iterations so we do not repeatedly acquire/release.
-struct PreparedEncode {
-    /// Image after decode + ops applied. Stored as `Arc` so the cheap path
-    /// (no ops, pre-decoded input) avoids a deep image clone.
-    processed: Arc<DynamicImage>,
-    final_color_state: ColorState,
-    input_format: Option<String>,
-    /// Whether the source originally had an EXIF segment. Detected once via
-    /// `has_exif` on the input bytes (not gated by `MAX_EXIF_SOURCE_BYTES`).
-    exif_present_in_source: bool,
-    decode_ms: f64,
-    ops_ms: f64,
-    usage_start: Option<ResourceUsage>,
-    input_size: u64,
-    start_total: Instant,
-    /// Hold the memory permit until the prepared image is dropped so the
-    /// reservation stays valid across multiple `encode_prepared` calls.
-    _permit_guard: memory::MemoryPermit,
-}
-
-/// Phase 1 of the pipeline: decode the input and apply all ops once.
-///
-/// Returns a [`PreparedEncode`] that can be fed to [`encode_prepared`] one or
-/// more times — useful for the byte-budget binary search where only the encode
-/// quality varies between iterations.
-fn prepare_for_encode(
-    source: Option<&Source>,
-    decoded: Option<&Arc<DynamicImage>>,
-    ops: &[Operation],
-    icc_present: bool,
-    auto_orient: bool,
-    firewall: &FirewallConfig,
-    format_for_memory_estimate: &OutputFormat,
-) -> std::result::Result<PreparedEncode, LazyImageError> {
-    let input_size = source.map(|s| s.len() as u64).unwrap_or(0);
-    let input_bytes = source.and_then(|s| s.as_bytes());
-    let input_format = input_bytes.and_then(detect_input_format);
-
-    // Memory backpressure: estimate before decode and acquire weighted permit.
-    // Use file-size-based fallback instead of hard-coded 100MB when header parsing fails.
-    let estimated_memory = source
-        .and_then(|s| s.as_bytes())
-        .and_then(|bytes| {
-            memory::estimate_memory_from_header(bytes, ops, Some(format_for_memory_estimate))
-        })
-        .unwrap_or_else(|| {
-            if source.is_none() {
-                return memory::ESTIMATED_MEMORY_PER_OPERATION;
-            }
-            let detected_format = source
-                .and_then(|s| s.as_bytes())
-                .and_then(crate::engine::decoder::detect_format);
-            memory::estimate_fallback_from_file_size(input_size, detected_format)
-        });
-    let source_reserved = source.map(|s| s.reserved_memory_bytes()).unwrap_or(0);
-    let sem = memory::memory_semaphore();
-    let total_memory = estimated_memory
-        .saturating_add(source_reserved)
-        .min(sem.capacity());
-    let permit = sem.acquire(total_memory);
-
-    let start_total = Instant::now();
-    let usage_start = get_resource_usage();
-    let mut stage_start = start_total;
-
-    // Pre-read orientation from EXIF header (before full decode)
-    let orientation = if auto_orient {
-        if let Some(bytes) = input_bytes {
-            // Enforce byte limit & metadata scan before EXIF parsing to honor firewall settings
-            firewall.enforce_source_len(bytes.len())?;
-            firewall.scan_metadata(bytes)?;
-            crate::engine::decoder::detect_exif_orientation(bytes)
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-
-    // EXIF presence detection: `has_exif` is not gated by `MAX_EXIF_SOURCE_BYTES`,
-    // so JPEGs larger than 8 MiB are still detected. Done here so the encode
-    // phase does not need to re-scan input bytes on every iteration.
-    let exif_present_in_source = input_bytes.map(has_exif).unwrap_or(false);
-
-    // 1. Decode
-    let img = decode_internal_from_parts(source, decoded, firewall)?;
-    firewall.enforce_timeout(start_total, "decode")?;
-    let decode_ms = stage_start.elapsed().as_secs_f64() * 1000.0;
-    stage_start = Instant::now();
-
-    // 2. Apply operations
-    let with_auto_orient;
-    let effective_ops = if let Some(o) = orientation {
-        // Clone only when we need to inject AutoOrient at the front.
-        with_auto_orient = {
-            let mut op_list = ops.to_vec();
-            op_list.insert(0, Operation::AutoOrient { orientation: o });
-            op_list
-        };
-        with_auto_orient.as_slice()
-    } else {
-        ops
-    };
-    let icc_state = if icc_present {
-        IccState::Present
-    } else {
-        IccState::Absent
-    };
-    let initial_state = ColorState::from_dynamic_image(&img, icc_state);
-    let tracked = apply_ops_tracked(img, effective_ops, initial_state)?;
-    let final_color_state = tracked.state;
-
-    // Materialize the processed image into an Arc:
-    // - Cow::Owned (ops applied OR decoded fresh from bytes) → wrap newly allocated buffer.
-    // - Cow::Borrowed (only possible when `decoded.is_some()` AND ops is effectively empty)
-    //   → clone the original Arc cheaply (atomic increment, no pixel copy).
-    let processed: Arc<DynamicImage> = match tracked.image {
-        Cow::Owned(img) => Arc::new(img),
-        Cow::Borrowed(_) => Arc::clone(
-            decoded
-                .expect("Cow::Borrowed implies decoded.is_some(); see decode_internal_from_parts"),
-        ),
-    };
-    firewall.enforce_timeout(start_total, "process")?;
-    let ops_ms = stage_start.elapsed().as_secs_f64() * 1000.0;
-
-    Ok(PreparedEncode {
-        processed,
-        final_color_state,
-        input_format,
-        exif_present_in_source,
-        decode_ms,
-        ops_ms,
-        usage_start,
-        input_size,
-        start_total,
-        _permit_guard: permit,
-    })
-}
-
-/// Phase 2 of the pipeline: encode the prepared image to the given format.
-///
-/// May be called multiple times against the same [`PreparedEncode`] — the
-/// byte-budget binary search uses this to vary only quality. Per-call timing
-/// (encode_ms, total_ms) is recorded into `metrics` if provided; decode_ms /
-/// ops_ms come from the cached prepare phase.
-fn encode_prepared(
-    prepared: &PreparedEncode,
-    ctx: &EncodeContext<'_>,
-    metrics: Option<&mut crate::ProcessingMetrics>,
-) -> std::result::Result<Vec<u8>, LazyImageError> {
-    let EncodeContext {
-        format,
-        icc_profile,
-        exif_data,
-        auto_orient,
-        policy,
-        firewall,
-        ..
-    } = *ctx;
-    let keep_icc = policy.effective_icc();
-    let keep_exif = policy.effective_exif();
-    let strip_gps = policy.strip_gps();
-
-    let encode_start = Instant::now();
-
-    // 3. Encode - only preserve ICC profile if keep_icc is true
-    let icc = if keep_icc {
-        icc_profile.map(|v| v.as_slice())
-    } else {
-        None // Strip metadata by default for security & smaller files
-    };
-
-    let processed: &DynamicImage = prepared.processed.as_ref();
-
-    // 4. Encode image to target format
-    let mut result = match format {
-        OutputFormat::Jpeg { quality, fast_mode } => {
-            encode_jpeg_with_settings(processed, *quality, icc, *fast_mode)
-        }
-        OutputFormat::Png => encode_png(processed, icc),
-        OutputFormat::WebP { quality } => encode_webp(processed, *quality, icc),
-        OutputFormat::Avif { quality } => encode_avif(processed, *quality, icc),
-    }?;
-
-    // 5. Embed EXIF metadata if requested.
-    if keep_exif {
-        if let Some(exif_data) = exif_data {
-            result = match format {
-                OutputFormat::Jpeg { .. } => {
-                    embed_exif_jpeg(result, exif_data.as_slice(), auto_orient, strip_gps)?
-                }
-                OutputFormat::Png => {
-                    embed_exif_png(result, exif_data.as_slice(), auto_orient, strip_gps)?
-                }
-                OutputFormat::WebP { .. } => {
-                    embed_exif_webp(result, exif_data.as_slice(), auto_orient, strip_gps)?
-                }
-                OutputFormat::Avif { .. } => result,
-            };
-        }
-    }
-    firewall.enforce_timeout(prepared.start_total, "encode")?;
-
-    let encode_ms = encode_start.elapsed().as_secs_f64() * 1000.0;
-
-    if let Some(m) = metrics {
-        // Get final resource usage & finalize metrics
-        let final_usage = get_resource_usage();
-        // Use tracked color state to reason about ICC preservation.
-        let icc_present = matches!(prepared.final_color_state.icc, IccState::Present);
-        let icc_preserved = keep_icc && icc_present;
-        // EXIF presence detection: prefer the in-memory copy that the API layer may
-        // have already extracted, otherwise fall back to a lightweight presence scan
-        // recorded during the prepare phase.
-        let exif_present = exif_data.is_some() || prepared.exif_present_in_source;
-        let exif_preserved =
-            keep_exif && exif_present && !matches!(format, OutputFormat::Avif { .. });
-        // GPS is only "stripped" if it was actually present in the source. When
-        // `keep_exif` is true, the API layer extracted EXIF eagerly, so we can
-        // parse it directly. When EXIF is being dropped by default policy,
-        // `exif_preserved` is false and `gps_stripped_from_exif` does not affect
-        // the final flag.
-        let gps_present_in_exif = exif_data
-            .map(|exif| exif_has_gps(exif.as_slice()))
-            .unwrap_or(false);
-        let gps_stripped_from_exif = exif_preserved && strip_gps && gps_present_in_exif;
-        let metadata_stripped = (icc_present && !icc_preserved)
-            || (exif_present && !exif_preserved)
-            || gps_stripped_from_exif;
-        let metadata_blocked_by_policy =
-            (keep_icc || keep_exif) && firewall.reject_metadata && (icc_present || exif_present);
-        let mut policy_violations = Vec::new();
-        if metadata_blocked_by_policy {
-            policy_violations.push("firewall_rejected_metadata".to_string());
-        }
-
-        m.decode_ms = prepared.decode_ms;
-        m.ops_ms = prepared.ops_ms;
-        m.encode_ms = encode_ms;
-        m.total_ms = prepared.start_total.elapsed().as_secs_f64() * 1000.0;
-        m.processing_time = m.total_ms / 1000.0;
-        m.version = PROCESSING_METRICS_VERSION.to_string();
-
-        if let (Some(start), Some(end)) = (prepared.usage_start.as_ref(), final_usage.as_ref()) {
-            m.cpu_time = (end.cpu_time - start.cpu_time).max(0.0);
-            m.peak_rss = end.memory_rss.min(u32::MAX as u64) as u32;
-        } else {
-            let (w, h) = processed.dimensions();
-            m.peak_rss =
-                ((w as u64 * h as u64 * 4) + result.len() as u64).min(u32::MAX as u64) as u32;
-        }
-
-        m.bytes_in = prepared.input_size.min(u32::MAX as u64) as u32;
-        m.bytes_out = (result.len() as u64).min(u32::MAX as u64) as u32;
-        m.compression_ratio = if m.bytes_in > 0 {
-            m.bytes_out as f64 / m.bytes_in as f64
-        } else {
-            0.0
-        };
-
-        m.format_in = prepared.input_format.clone();
-        m.format_out = format.as_str().to_string();
-        m.icc_preserved = icc_preserved;
-        m.metadata_stripped = metadata_stripped;
-        m.policy_violations = policy_violations;
-    }
-
-    Ok(result)
-}
-
-fn process_and_encode_from_parts(
-    ctx: &EncodeContext<'_>,
-    metrics: Option<&mut crate::ProcessingMetrics>,
-) -> std::result::Result<Vec<u8>, LazyImageError> {
-    let prepared = prepare_for_encode(
-        ctx.source,
-        ctx.decoded,
-        ctx.ops,
-        ctx.icc_present,
-        ctx.auto_orient,
-        ctx.firewall,
-        ctx.format,
-    )?;
-    encode_prepared(&prepared, ctx, metrics)
-}
-
-impl TaskContext {
-    /// Build a borrowed [`EncodeContext`] view over this task's fields.
-    ///
-    /// Every per-task field (source, ops, format, ICC/EXIF, policy, firewall)
-    /// is exposed as a borrow; the resulting context can be passed by reference
-    /// to [`process_and_encode_from_parts`] and [`encode_prepared`].
-    fn encode_context(&self) -> EncodeContext<'_> {
-        EncodeContext {
-            source: self.source.as_ref(),
-            decoded: self.decoded.as_ref(),
-            ops: &self.ops,
-            format: &self.format,
-            icc_profile: self.icc_profile.as_ref(),
-            icc_present: self.icc_present,
-            exif_data: self.exif_data.as_ref(),
-            auto_orient: self.auto_orient,
-            policy: &self.metadata_policy,
-            firewall: &self.firewall,
-        }
-    }
-
-    /// Process image: decode → apply ops → encode
-    /// This is the core processing pipeline shared by toBuffer and toFile.
-    /// Returns LazyImageError directly (not wrapped in napi::Error) so that
-    /// Task::reject can properly create error objects with code/category.
-    pub(crate) fn process_and_encode(
-        &self,
-        metrics: Option<&mut crate::ProcessingMetrics>,
-    ) -> std::result::Result<Vec<u8>, LazyImageError> {
-        process_and_encode_from_parts(&self.encode_context(), metrics)
-    }
-
-    /// Store an error for later use in reject(), returning a napi::Error for compute().
-    #[cfg(feature = "napi")]
-    fn store_and_convert_error(&mut self, lazy_err: LazyImageError) -> napi::Error {
-        self.last_error = Some(lazy_err.clone());
-        napi::Error::from(lazy_err)
-    }
-
-    /// Retrieve stored error or create a fallback from the napi::Error message.
-    #[cfg(feature = "napi")]
-    fn take_stored_error(&mut self, err: napi::Error) -> LazyImageError {
-        self.last_error
-            .take()
-            .unwrap_or_else(|| LazyImageError::generic(err.to_string()))
-    }
-}
-
-// Test-only helper for decode assertions in engine.rs tests.
-#[cfg(test)]
-impl TaskContext {
-    pub(crate) fn decode(&self) -> std::result::Result<Cow<'_, DynamicImage>, LazyImageError> {
-        decode_internal_from_parts(self.source.as_ref(), self.decoded.as_ref(), &self.firewall)
-    }
-}
-
-// ============================================================================================
-// Macro: define_encode_task!
-//
-// Generates a task struct containing `pub ctx: TaskContext` plus any additional
-// fields, plus the NAPI `Task` trait implementation with customizable
-// compute / resolve / reject bodies.
-// ============================================================================================
-
-/// Generate a task struct with `ctx: TaskContext` and optional extra fields,
-/// plus its NAPI `Task` impl.
-///
-/// Syntax:
-/// ```ignore
-/// define_encode_task! {
-///     /// doc comment
-///     pub struct MyTask {
-///         // extra fields (TaskContext is always included as `ctx`)
-///         pub extra_field: Type,
-///     }
-///     napi {
-///         type Output = OutputType;
-///         type JsValue = JsType;
-///         compute(self) { ... }
-///         resolve(self, env, output) { ... }
-///     }
-/// }
-/// ```
-macro_rules! define_encode_task {
-    (
-        $( #[$meta:meta] )*
-        pub struct $name:ident {
-            $( $( #[$fmeta:meta] )* pub $field:ident : $fty:ty ),* $(,)?
-        }
-        napi {
-            type Output = $out:ty;
-            type JsValue = $jsval:ty;
-            compute($self_compute:ident) $compute_body:block
-            resolve($self_resolve:ident, $env_resolve:ident, $output_resolve:ident) $resolve_body:block
-        }
-    ) => {
-        $( #[$meta] )*
-        pub struct $name {
-            pub ctx: TaskContext,
-            $( $( #[$fmeta] )* pub $field : $fty, )*
-        }
-
-        // Compile-time `Send` guarantee. NAPI's `AsyncTask` moves task structs
-        // onto a worker thread for `compute()`, so any non-Send field would
-        // surface as a runtime error rather than at build time. This assertion
-        // catches it during compilation, before the napi attribute even
-        // expands.
-        const _: fn() = || {
-            fn assert_send<T: Send>() {}
-            assert_send::<$name>();
-        };
-
-        #[cfg(feature = "napi")]
-        #[napi]
-        impl Task for $name {
-            type Output = $out;
-            type JsValue = $jsval;
-
-            fn compute(&mut $self_compute) -> Result<Self::Output> $compute_body
-
-            fn resolve(&mut $self_resolve, $env_resolve: Env, $output_resolve: Self::Output) -> Result<Self::JsValue> $resolve_body
-
-            fn reject(&mut self, env: Env, err: napi::Error) -> Result<Self::JsValue> {
-                let lazy_err = self.ctx.take_stored_error(err);
-                let napi_err = crate::error::napi_error_with_code(&env, lazy_err)?;
-                Err(napi_err)
-            }
-        }
-    };
-}
-
-// ============================================================================================
-// Task definitions using the macro
-// ============================================================================================
-
-define_encode_task! {
-    pub struct EncodeTask {}
-    napi {
-        type Output = Vec<u8>;
-        type JsValue = napi::bindgen_prelude::Buffer;
-        compute(self) {
-            match self.ctx.process_and_encode(None) {
-                Ok(result) => {
-                    #[cfg(feature = "napi")]
-                    { self.ctx.last_error = None; }
-                    Ok(result)
-                }
-                Err(lazy_err) => {
-                    Err(self.ctx.store_and_convert_error(lazy_err))
-                }
-            }
-        }
-        resolve(self, env, output) {
-            BufferSlice::from_data(&env, output)?.into_buffer(&env)
-        }
-    }
-}
-
-#[cfg(test)]
-impl EncodeTask {
-    /// Delegate to TaskContext for backward compatibility with test callers.
-    pub(crate) fn process_and_encode(
-        &self,
-        metrics: Option<&mut crate::ProcessingMetrics>,
-    ) -> std::result::Result<Vec<u8>, LazyImageError> {
-        self.ctx.process_and_encode(metrics)
-    }
-
-    pub(crate) fn decode(&self) -> std::result::Result<Cow<'_, DynamicImage>, LazyImageError> {
-        self.ctx.decode()
-    }
-}
-
-define_encode_task! {
-    pub struct EncodeWithMetricsTask {}
-    napi {
-        type Output = (Vec<u8>, crate::ProcessingMetrics);
-        type JsValue = crate::OutputWithMetrics;
-        compute(self) {
-            let mut metrics = crate::ProcessingMetrics::default();
-            match self.ctx.process_and_encode(Some(&mut metrics)) {
-                Ok(data) => {
-                    #[cfg(feature = "napi")]
-                    { self.ctx.last_error = None; }
-                    Ok((data, metrics))
-                }
-                Err(lazy_err) => {
-                    Err(self.ctx.store_and_convert_error(lazy_err))
-                }
-            }
-        }
-        resolve(self, env, output) {
-            let (data, metrics) = output;
-            let js_buffer = BufferSlice::from_data(&env, data)?.into_buffer(&env)?;
-            Ok(crate::OutputWithMetrics {
-                data: js_buffer,
-                metrics,
-            })
-        }
-    }
-}
-
-/// Result payload for the byte-budget binary search (before NAPI conversion).
-#[derive(Debug)]
-pub struct TargetBytesOutput {
-    pub data: Vec<u8>,
-    pub quality: u8,
-    pub bytes_out: u32,
-    pub budget_met: bool,
-    pub target_bytes: u32,
-    pub metrics: crate::ProcessingMetrics,
-}
-
-define_encode_task! {
-    /// Async task that performs byte-budget binary search entirely in Rust.
-    ///
-    /// Previous implementation did this in JS, requiring ~7 NAPI round-trips per
-    /// call (one per binary-search iteration). Moving the loop to Rust eliminates
-    /// all intermediate JS↔NAPI boundary crossings.
-    pub struct EncodeTargetBytesTask {
-        pub target_bytes: u32,
-        pub min_quality: u8,
-        pub max_quality: u8,
-        pub quality_floor_policy_strict: bool,
-    }
-    napi {
-        type Output = TargetBytesOutput;
-        type JsValue = crate::TargetBytesResult;
-        compute(self) {
-            match self.search() {
-                Ok(result) => {
-                    #[cfg(feature = "napi")]
-                    { self.ctx.last_error = None; }
-                    Ok(result)
-                }
-                Err(lazy_err) => {
-                    Err(self.ctx.store_and_convert_error(lazy_err))
-                }
-            }
-        }
-        resolve(self, env, output) {
-            let js_buffer = BufferSlice::from_data(&env, output.data)?.into_buffer(&env)?;
-            Ok(crate::TargetBytesResult {
-                data: js_buffer,
-                quality: output.quality as u32,
-                bytes_out: output.bytes_out,
-                budget_met: output.budget_met,
-                target_bytes: output.target_bytes,
-                metrics: output.metrics,
-            })
-        }
-    }
-}
-
-impl EncodeTargetBytesTask {
-    /// Run binary search over quality to find the best encode within byte budget.
-    ///
-    /// Decode + ops are performed **once** via [`prepare_for_encode`]; only the
-    /// encode step (and EXIF embed) re-runs per iteration. For a 4000x3000 JPEG
-    /// with a resize op this eliminates a 5-7x re-decode penalty (issue #558).
-    fn search(&self) -> std::result::Result<TargetBytesOutput, LazyImageError> {
-        // Decode + apply ops once. Memory permit is held by `prepared` so the
-        // semaphore reservation stays valid across every encode iteration below.
-        let base_ctx = self.ctx.encode_context();
-        let prepared = prepare_for_encode(
-            base_ctx.source,
-            base_ctx.decoded,
-            base_ctx.ops,
-            base_ctx.icc_present,
-            base_ctx.auto_orient,
-            base_ctx.firewall,
-            base_ctx.format,
-        )?;
-
-        let mut low = self.min_quality;
-        let mut high = self.max_quality;
-        let mut best_under: Option<(Vec<u8>, u8, u32, crate::ProcessingMetrics)> = None;
-        let mut best_over: Option<(Vec<u8>, u8, u32, crate::ProcessingMetrics)> = None;
-
-        while low <= high {
-            let quality = low + (high - low) / 2;
-            let format_at_q = self.ctx.format.with_quality(quality);
-            // Reuse the base context, overriding only the per-iteration format.
-            let iter_ctx = EncodeContext {
-                format: &format_at_q,
-                ..base_ctx
-            };
-
-            let mut metrics = crate::ProcessingMetrics::default();
-            let encoded = encode_prepared(&prepared, &iter_ctx, Some(&mut metrics))?;
-
-            let size = encoded.len() as u32;
-
-            if size <= self.target_bytes {
-                best_under = Some((encoded, quality, size, metrics));
-                if quality == u8::MAX {
-                    break;
-                }
-                low = quality + 1;
-            } else {
-                let dominated = best_over.as_ref().is_some_and(|(_, _, prev_size, _)| {
-                    size > *prev_size
-                        || (size == *prev_size && quality <= best_over.as_ref().unwrap().1)
-                });
-                if best_over.is_none() || !dominated {
-                    best_over = Some((encoded, quality, size, metrics));
-                }
-                high = quality.saturating_sub(1);
-                if quality == 0 {
-                    break;
-                }
-            }
-        }
-
-        let (data, quality, bytes_out, metrics) = match (best_under, best_over) {
-            (Some(under), _) => under,
-            (None, Some(over)) => over,
-            (None, None) => {
-                return Err(LazyImageError::encode_failed(
-                    self.ctx.format.as_str(),
-                    "Failed to encode any candidate while searching for target bytes",
-                ));
-            }
-        };
-
-        let budget_met = bytes_out <= self.target_bytes;
-        if !budget_met && self.quality_floor_policy_strict {
-            return Err(LazyImageError::encode_failed(
-                self.ctx.format.as_str(),
-                format!(
-                    "Unable to meet targetBytes={} within quality range {}-{}",
-                    self.target_bytes, self.min_quality, self.max_quality,
-                ),
-            ));
-        }
-
-        Ok(TargetBytesOutput {
-            data,
-            quality,
-            bytes_out,
-            budget_met,
-            target_bytes: self.target_bytes,
-            metrics,
-        })
-    }
-}
-
-/// Write `data` to `output_path` atomically (temp-file + fsync + rename).
-///
-/// All file outputs go through this single helper to guarantee crash-safe
-/// writes: a power loss or process kill mid-write leaves the destination
-/// either untouched or fully written, never half-populated. The temp file
-/// lives in the destination's parent directory so the final `persist()`
-/// is a same-filesystem rename (atomic per POSIX).
-///
-/// Returns the number of bytes written as `u32`. The check fails for files
-/// larger than 4 GiB, but no realistic image encode produces that much
-/// output; callers that don't need the count simply discard the return value.
-fn write_encoded_output(
-    output_path: &str,
-    data: &[u8],
-) -> std::result::Result<u32, LazyImageError> {
-    use std::io::Write;
-    use tempfile::NamedTempFile;
-
-    let output_dir = std::path::Path::new(output_path).parent().ok_or_else(|| {
-        LazyImageError::invalid_argument(
-            "path",
-            output_path.to_string(),
-            "output path must include a parent directory",
-        )
-    })?;
-
-    let mut temp_file = NamedTempFile::new_in(output_dir).map_err(|e| {
-        LazyImageError::file_write_failed(output_dir.to_string_lossy().to_string(), e)
-    })?;
-
-    let temp_path = temp_file.path().to_path_buf();
-    let bytes_written = data
-        .len()
-        .try_into()
-        .map_err(|_| LazyImageError::internal_panic("file size exceeds 4GB limit (u32::MAX)"))?;
-    temp_file
-        .write_all(data)
-        .map_err(|e| LazyImageError::file_write_failed(temp_path.display().to_string(), e))?;
-    temp_file
-        .as_file_mut()
-        .sync_all()
-        .map_err(|e| LazyImageError::file_write_failed(temp_path.display().to_string(), e))?;
-
-    temp_file.persist(output_path).map_err(|e| {
-        let io_error = std::io::Error::other(format!("failed to persist file: {e}"));
-        LazyImageError::file_write_failed(output_path.to_string(), io_error)
-    })?;
-
-    Ok(bytes_written)
-}
-
-define_encode_task! {
-    pub struct WriteFileTask {
-        pub output_path: String,
-    }
-    napi {
-        type Output = u32;
-        type JsValue = u32;
-        compute(self) {
-            let data = match self.ctx.process_and_encode(None) {
-                Ok(data) => data,
-                Err(lazy_err) => {
-                    return Err(self.ctx.store_and_convert_error(lazy_err));
-                }
-            };
-
-            match write_encoded_output(&self.output_path, &data) {
-                Ok(bytes_written) => Ok(bytes_written),
-                Err(lazy_err) => {
-                    Err(self.ctx.store_and_convert_error(lazy_err))
-                }
-            }
-        }
-        resolve(self, _env, output) {
-            Ok(output)
-        }
-    }
-}
-
-define_encode_task! {
-    pub struct WriteFileWithMetricsTask {
-        pub output_path: String,
-    }
-    napi {
-        type Output = (u32, crate::ProcessingMetrics);
-        type JsValue = crate::FileOutputWithMetrics;
-        compute(self) {
-            let mut metrics = crate::ProcessingMetrics::default();
-            let data = match self.ctx.process_and_encode(Some(&mut metrics)) {
-                Ok(data) => data,
-                Err(lazy_err) => {
-                    return Err(self.ctx.store_and_convert_error(lazy_err));
-                }
-            };
-
-            match write_encoded_output(&self.output_path, &data) {
-                Ok(bytes_written) => Ok((bytes_written, metrics)),
-                Err(lazy_err) => {
-                    Err(self.ctx.store_and_convert_error(lazy_err))
-                }
-            }
-        }
-        resolve(self, _env, output) {
-            let (bytes_written, metrics) = output;
-            Ok(crate::FileOutputWithMetrics {
-                bytes_written,
-                metrics,
-            })
-        }
-    }
-}
-
-// ============================================================================================
-// Unified encode tasks — power encode() and encodeToFile()
-// ============================================================================================
-
-define_encode_task! {
-    /// Task for `encode()` — returns `EncodeResult { data, metrics? }`.
-    pub struct UnifiedEncodeTask {
-        pub want_metrics: bool,
-    }
-    napi {
-        type Output = (Vec<u8>, Option<crate::ProcessingMetrics>);
-        type JsValue = crate::EncodeResult;
-        compute(self) {
-            let mut metrics = if self.want_metrics {
-                Some(crate::ProcessingMetrics::default())
-            } else {
-                None
-            };
-
-            match self.ctx.process_and_encode(metrics.as_mut()) {
-                Ok(data) => {
-                    #[cfg(feature = "napi")]
-                    { self.ctx.last_error = None; }
-                    Ok((data, metrics))
-                }
-                Err(lazy_err) => {
-                    Err(self.ctx.store_and_convert_error(lazy_err))
-                }
-            }
-        }
-        resolve(self, env, output) {
-            let (data, metrics) = output;
-            let js_buffer = BufferSlice::from_data(&env, data)?.into_buffer(&env)?;
-            Ok(crate::EncodeResult {
-                data: js_buffer,
-                metrics,
-            })
-        }
-    }
-}
-
-define_encode_task! {
-    /// Task for `encodeToFile()` — returns `FileEncodeResult { bytesWritten, metrics? }`.
-    pub struct UnifiedEncodeToFileTask {
-        pub want_metrics: bool,
-        pub output_path: String,
-    }
-    napi {
-        type Output = (u32, Option<crate::ProcessingMetrics>);
-        type JsValue = crate::FileEncodeResult;
-        compute(self) {
-            let mut metrics = if self.want_metrics {
-                Some(crate::ProcessingMetrics::default())
-            } else {
-                None
-            };
-
-            let data = match self.ctx.process_and_encode(metrics.as_mut()) {
-                Ok(d) => d,
-                Err(lazy_err) => {
-                    return Err(self.ctx.store_and_convert_error(lazy_err));
-                }
-            };
-
-            match write_encoded_output(&self.output_path, &data) {
-                Ok(bytes_written) => {
-                    #[cfg(feature = "napi")]
-                    { self.ctx.last_error = None; }
-                    Ok((bytes_written, metrics))
-                }
-                Err(lazy_err) => {
-                    Err(self.ctx.store_and_convert_error(lazy_err))
-                }
-            }
-        }
-        resolve(self, _env, output) {
-            let (bytes_written, metrics) = output;
-            Ok(crate::FileEncodeResult {
-                bytes_written,
-                metrics,
-            })
-        }
-    }
-}
-
-// ============================================================================================
-// Batch tasks — kept manual due to structurally different parallelism (rayon)
-// ============================================================================================
-
-struct BatchFileSuccess {
-    output_path: String,
-    metrics: Option<crate::ProcessingMetrics>,
-}
-
-/// Borrowed parameter bundle for [`process_batch_file`].
-///
-/// Holds the per-batch configuration (output directory, ops, format, policy,
-/// firewall, etc.) that is identical for every file in a batch. Built once
-/// outside the rayon loop and shared (by `Copy`) across every worker closure.
-#[derive(Clone, Copy)]
-struct BatchFileContext<'a> {
-    output_dir: &'a str,
-    ops: &'a [Operation],
-    format: &'a OutputFormat,
-    auto_orient: bool,
-    policy: &'a MetadataPolicy,
-    firewall: &'a FirewallConfig,
-    collect_metrics: bool,
-}
-
-fn process_batch_file(
-    input_path: &str,
-    ctx: &BatchFileContext<'_>,
-) -> std::result::Result<BatchFileSuccess, LazyImageError> {
-    let keep_icc = ctx.policy.effective_icc();
-    let keep_exif = ctx.policy.effective_exif();
-    use std::path::Path;
-    use std::sync::Arc;
-
-    let input_len = std::fs::metadata(input_path)
-        .map_err(|e| {
-            if e.kind() == std::io::ErrorKind::NotFound {
-                LazyImageError::file_not_found(input_path.to_string())
-            } else {
-                LazyImageError::file_read_failed(input_path.to_string(), e)
-            }
-        })?
-        .len() as usize;
-    ctx.firewall.enforce_source_len(input_len)?;
-
-    let source = crate::engine::io::load_file_safe(Path::new(input_path))?;
-    let data = source.as_bytes().expect("source always has bytes");
-
-    ctx.firewall.scan_metadata(data)?;
-
-    let extracted_icc = extract_icc_profile(data)?;
-    let icc_present = extracted_icc.is_some();
-    let icc_profile = if keep_icc {
-        extracted_icc.map(Arc::new)
-    } else {
-        None
-    };
-    let exif_data = if keep_exif {
-        extract_exif_raw(data).map(Arc::new)
-    } else {
-        None
-    };
-
-    let mut metrics = ctx.collect_metrics.then(crate::ProcessingMetrics::default);
-    let encode_ctx = EncodeContext {
-        source: Some(&source),
-        decoded: None,
-        ops: ctx.ops,
-        format: ctx.format,
-        icc_profile: icc_profile.as_ref(),
-        icc_present,
-        exif_data: exif_data.as_ref(),
-        auto_orient: ctx.auto_orient,
-        policy: ctx.policy,
-        firewall: ctx.firewall,
-    };
-    let encoded = process_and_encode_from_parts(&encode_ctx, metrics.as_mut())?;
-
-    let filename = Path::new(input_path)
-        .file_name()
-        .ok_or_else(|| LazyImageError::internal_panic("invalid filename"))?;
-    let extension = match ctx.format {
-        OutputFormat::Jpeg { .. } => "jpg",
-        OutputFormat::Png => "png",
-        OutputFormat::WebP { .. } => "webp",
-        OutputFormat::Avif { .. } => "avif",
-    };
-    let output_filename = Path::new(filename).with_extension(extension);
-    let output_path = Path::new(ctx.output_dir).join(output_filename);
-    write_encoded_output(output_path.to_string_lossy().as_ref(), &encoded)?;
-
-    Ok(BatchFileSuccess {
-        output_path: output_path.to_string_lossy().to_string(),
-        metrics,
-    })
-}
-
-pub struct BatchTask {
-    pub inputs: Vec<String>,
-    pub output_dir: String,
-    pub ops: Vec<Operation>,
-    pub format: OutputFormat,
-    pub concurrency: u32,
-    pub auto_orient: bool,
-    /// Single source of truth for metadata preservation decisions
-    pub metadata_policy: MetadataPolicy,
-    pub firewall: FirewallConfig,
-    /// Last error that occurred during compute (for use in reject)
-    #[cfg(feature = "napi")]
-    pub(crate) last_error: Option<LazyImageError>,
-}
-
-pub struct BatchWithMetricsTask {
-    pub inputs: Vec<String>,
-    pub output_dir: String,
-    pub ops: Vec<Operation>,
-    pub format: OutputFormat,
-    pub concurrency: u32,
-    pub auto_orient: bool,
-    /// Single source of truth for metadata preservation decisions
-    pub metadata_policy: MetadataPolicy,
-    pub firewall: FirewallConfig,
-    #[cfg(feature = "napi")]
-    pub(crate) last_error: Option<LazyImageError>,
-}
-
-// Compile-time `Send` assertions for the manually-implemented batch tasks.
-// Matches the assertion the `define_encode_task!` macro emits for the other
-// task types — see the comment near the macro definition.
-const _: fn() = || {
-    fn assert_send<T: Send>() {}
-    assert_send::<BatchTask>();
-    assert_send::<BatchWithMetricsTask>();
-};
-
-#[cfg(feature = "napi")]
-#[napi]
-impl Task for BatchTask {
-    type Output = Vec<BatchResult>;
-    type JsValue = Vec<BatchResult>;
-
-    fn compute(&mut self) -> Result<Self::Output> {
-        use std::fs;
-        use std::path::Path;
-
-        if !Path::new(&self.output_dir).exists() {
-            fs::create_dir_all(&self.output_dir).map_err(|e| {
-                let lazy_err = LazyImageError::file_write_failed(self.output_dir.clone(), e);
-                self.last_error = Some(lazy_err.clone());
-                napi::Error::from(lazy_err)
-            })?;
-        }
-
-        // concurrency = 0 means "use default" (auto-detected via CPU and memory)
-        // concurrency = 1..MAX_CONCURRENCY means "use specified number of concurrent operations"
-        if self.concurrency > pool::MAX_CONCURRENCY as u32 {
-            let lazy_err = LazyImageError::invalid_argument(
-                "concurrency",
-                self.concurrency.to_string(),
-                format!("must be 0 or 1-{}", pool::MAX_CONCURRENCY),
-            );
-            self.last_error = Some(lazy_err.clone());
-            return Err(napi::Error::from(lazy_err));
-        }
-
-        let effective_concurrency = pool::resolve_effective_concurrency(self.concurrency);
-        let effective_concurrency_u32 = u32::try_from(effective_concurrency).ok();
-        let auto_concurrency = Some(self.concurrency == 0);
-
-        // Helper closure to process a single image
-        let firewall = self.firewall.clone();
-        let batch_ctx = BatchFileContext {
-            output_dir: &self.output_dir,
-            ops: &self.ops,
-            format: &self.format,
-            auto_orient: self.auto_orient,
-            policy: &self.metadata_policy,
-            firewall: &firewall,
-            collect_metrics: false,
-        };
-        let process_one = |input_path: &String| -> BatchResult {
-            let result = process_batch_file(input_path, &batch_ctx);
-
-            match result {
-                Ok(output) => BatchResult {
-                    source: input_path.clone(),
-                    success: true,
-                    error: None,
-                    output_path: Some(output.output_path),
-                    error_code: None,
-                    error_category: None,
-                    effective_concurrency: effective_concurrency_u32,
-                    auto_concurrency,
-                },
-                Err(err) => {
-                    let error_code = err.code();
-                    let error_msg = format!("[{}] {}: {}", error_code.as_str(), input_path, err);
-                    let category = error_code.category();
-                    BatchResult {
-                        source: input_path.clone(),
-                        success: false,
-                        error: Some(error_msg),
-                        output_path: None,
-                        error_code: Some(error_code.as_str().to_string()),
-                        error_category: Some(category),
-                        effective_concurrency: effective_concurrency_u32,
-                        auto_concurrency,
-                    }
-                }
-            }
-        };
-
-        // Use the shared rayon pool to avoid per-call pool construction overhead.
-        // Memory backpressure is still handled by WeightedSemaphore.
-        //
-        // When effective_concurrency < inputs.len(), run exactly that many worker
-        // tasks on the global pool and pull work dynamically via an atomic index.
-        // This preserves the concurrency cap while avoiding chunk-induced tail latency.
-        let pool = match pool::get_pool() {
-            Ok(pool) => pool,
-            Err(err) => {
-                self.last_error = Some(err.clone());
-                return Err(napi::Error::from(err));
-            }
-        };
-
-        let results: Vec<BatchResult> = if effective_concurrency >= self.inputs.len() {
-            pool.install(|| self.inputs.par_iter().map(process_one).collect())
-        } else {
-            use std::sync::atomic::{AtomicUsize, Ordering};
-            use std::sync::{Arc, Mutex};
-
-            let next_index = Arc::new(AtomicUsize::new(0));
-            let indexed_results = Arc::new(Mutex::new(Vec::<(usize, BatchResult)>::with_capacity(
-                self.inputs.len(),
-            )));
-
-            pool.install(|| {
-                rayon::scope(|scope| {
-                    for _ in 0..effective_concurrency {
-                        let next_index = Arc::clone(&next_index);
-                        let indexed_results = Arc::clone(&indexed_results);
-                        let inputs = &self.inputs;
-                        scope.spawn(move |_| loop {
-                            let idx = next_index.fetch_add(1, Ordering::Relaxed);
-                            if idx >= inputs.len() {
-                                break;
-                            }
-
-                            let result = process_one(&inputs[idx]);
-                            let mut guard = indexed_results
-                                .lock()
-                                .unwrap_or_else(|poisoned| poisoned.into_inner());
-                            guard.push((idx, result));
-                        });
-                    }
-                });
-            });
-
-            let mut indexed_results = {
-                let mut guard = indexed_results
-                    .lock()
-                    .unwrap_or_else(|poisoned| poisoned.into_inner());
-                std::mem::take(&mut *guard)
-            };
-            indexed_results.sort_unstable_by_key(|(idx, _)| *idx);
-            indexed_results
-                .into_iter()
-                .map(|(_, result)| result)
-                .collect()
-        };
-
-        Ok(results)
-    }
-
-    fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
-        Ok(output)
-    }
-
-    fn reject(&mut self, env: Env, err: napi::Error) -> Result<Self::JsValue> {
-        // Use stored error if available, otherwise try to extract from napi::Error
-        let lazy_err = self.last_error.take().unwrap_or_else(|| {
-            // Fallback: create a generic error from the napi::Error message
-            // This should not happen if all error paths properly preserve LazyImageError
-            LazyImageError::generic(err.to_string())
-        });
-        let napi_err = crate::error::napi_error_with_code(&env, lazy_err)?;
-        Err(napi_err)
-    }
-}
-
-#[cfg(feature = "napi")]
-#[napi]
-impl Task for BatchWithMetricsTask {
-    type Output = crate::BatchOutputWithMetrics;
-    type JsValue = crate::BatchOutputWithMetrics;
-
-    fn compute(&mut self) -> Result<Self::Output> {
-        use std::fs;
-        use std::path::Path;
-
-        if !Path::new(&self.output_dir).exists() {
-            fs::create_dir_all(&self.output_dir).map_err(|e| {
-                let lazy_err = LazyImageError::file_write_failed(self.output_dir.clone(), e);
-                self.last_error = Some(lazy_err.clone());
-                napi::Error::from(lazy_err)
-            })?;
-        }
-
-        if self.concurrency > pool::MAX_CONCURRENCY as u32 {
-            let lazy_err = LazyImageError::invalid_argument(
-                "concurrency",
-                self.concurrency.to_string(),
-                format!("must be 0 or 1-{}", pool::MAX_CONCURRENCY),
-            );
-            self.last_error = Some(lazy_err.clone());
-            return Err(napi::Error::from(lazy_err));
-        }
-
-        let effective_concurrency = pool::resolve_effective_concurrency(self.concurrency);
-        let effective_concurrency_u32 = u32::try_from(effective_concurrency).unwrap_or(u32::MAX);
-        let auto_concurrency = self.concurrency == 0;
-
-        let firewall = self.firewall.clone();
-        let batch_ctx = BatchFileContext {
-            output_dir: &self.output_dir,
-            ops: &self.ops,
-            format: &self.format,
-            auto_orient: self.auto_orient,
-            policy: &self.metadata_policy,
-            firewall: &firewall,
-            collect_metrics: true,
-        };
-
-        let process_one = |input_path: &String| -> crate::BatchResultWithMetrics {
-            match process_batch_file(input_path, &batch_ctx) {
-                Ok(output) => crate::BatchResultWithMetrics {
-                    source: input_path.clone(),
-                    success: true,
-                    error: None,
-                    output_path: Some(output.output_path),
-                    error_code: None,
-                    error_category: None,
-                    effective_concurrency: Some(effective_concurrency_u32),
-                    auto_concurrency: Some(auto_concurrency),
-                    metrics: output.metrics,
-                },
-                Err(err) => {
-                    let error_code = err.code();
-                    let error_msg = format!("[{}] {}: {}", error_code.as_str(), input_path, err);
-                    let category = error_code.category();
-                    crate::BatchResultWithMetrics {
-                        source: input_path.clone(),
-                        success: false,
-                        error: Some(error_msg),
-                        output_path: None,
-                        error_code: Some(error_code.as_str().to_string()),
-                        error_category: Some(category),
-                        effective_concurrency: Some(effective_concurrency_u32),
-                        auto_concurrency: Some(auto_concurrency),
-                        metrics: None,
-                    }
-                }
-            }
-        };
-
-        let pool = match pool::get_pool() {
-            Ok(pool) => pool,
-            Err(err) => {
-                self.last_error = Some(err.clone());
-                return Err(napi::Error::from(err));
-            }
-        };
-
-        let items: Vec<crate::BatchResultWithMetrics> =
-            if effective_concurrency >= self.inputs.len() {
-                pool.install(|| self.inputs.par_iter().map(process_one).collect())
-            } else {
-                use std::sync::atomic::{AtomicUsize, Ordering};
-                use std::sync::{Arc, Mutex};
-
-                let inputs = Arc::new(&self.inputs);
-                let out = Arc::new(Mutex::new(Vec::with_capacity(self.inputs.len())));
-                let next = AtomicUsize::new(0);
-
-                pool.install(|| {
-                    rayon::scope(|s| {
-                        for _ in 0..effective_concurrency {
-                            let inputs = Arc::clone(&inputs);
-                            let out = Arc::clone(&out);
-                            let next_ref = &next;
-                            s.spawn(move |_| loop {
-                                let idx = next_ref.fetch_add(1, Ordering::Relaxed);
-                                if idx >= inputs.len() {
-                                    break;
-                                }
-                                let item = process_one(&inputs[idx]);
-                                out.lock().unwrap().push((idx, item));
-                            });
-                        }
-                    });
-                });
-
-                let mut pairs = Arc::try_unwrap(out).unwrap().into_inner().unwrap();
-                pairs.sort_by_key(|(idx, _)| *idx);
-                pairs.into_iter().map(|(_, item)| item).collect()
-            };
-
-        let mut successful_items = 0u32;
-        let mut failed_items = 0u32;
-        let mut total_bytes_in = 0f64;
-        let mut total_bytes_out = 0f64;
-        let mut total_decode_ms = 0f64;
-        let mut total_ops_ms = 0f64;
-        let mut total_encode_ms = 0f64;
-        let mut total_cpu_time = 0f64;
-        let mut total_wall_ms = 0f64;
-
-        for item in &items {
-            if item.success {
-                successful_items += 1;
-            } else {
-                failed_items += 1;
-            }
-            if let Some(metrics) = item.metrics.as_ref() {
-                total_bytes_in += metrics.bytes_in as f64;
-                total_bytes_out += metrics.bytes_out as f64;
-                total_decode_ms += metrics.decode_ms;
-                total_ops_ms += metrics.ops_ms;
-                total_encode_ms += metrics.encode_ms;
-                total_cpu_time += metrics.cpu_time;
-                total_wall_ms += metrics.total_ms;
-            }
-        }
-
-        Ok(crate::BatchOutputWithMetrics {
-            items,
-            summary: crate::BatchMetricsSummary {
-                total_items: self.inputs.len() as u32,
-                successful_items,
-                failed_items,
-                effective_concurrency: effective_concurrency_u32,
-                auto_concurrency,
-                total_bytes_in,
-                total_bytes_out,
-                total_decode_ms,
-                total_ops_ms,
-                total_encode_ms,
-                total_cpu_time,
-                total_wall_ms,
-            },
-        })
-    }
-
-    fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
-        Ok(output)
-    }
-
-    fn reject(&mut self, env: Env, err: napi::Error) -> Result<Self::JsValue> {
-        let lazy_err = self
-            .last_error
-            .take()
-            .unwrap_or_else(|| LazyImageError::generic(err.to_string()));
-        let napi_err = crate::error::napi_error_with_code(&env, lazy_err)?;
-        Err(napi_err)
-    }
-}
-
-// ============================================================================================
-// Tests for coverage when NAPI is disabled (CI runs coverage with --no-default-features)
-// ============================================================================================
-#[cfg(all(test, not(feature = "napi")))]
-mod non_napi_tests {
-    use super::*;
-    use crate::engine::firewall::FirewallConfig;
-    use crate::engine::io::Source;
-    use crate::ops::ResizeFit;
-    use image::{ImageBuffer, ImageFormat, Rgba};
-
-    fn sample_png_bytes() -> Vec<u8> {
-        let img: ImageBuffer<Rgba<u8>, Vec<u8>> =
-            ImageBuffer::from_pixel(4, 4, Rgba([10, 20, 30, 255]));
-        let mut bytes = Vec::new();
-        img.write_to(&mut std::io::Cursor::new(&mut bytes), ImageFormat::Png)
-            .unwrap();
-        bytes
-    }
-
-    /// Helper to create a TaskContext with a pre-decoded image.
-    fn make_ctx_with_decoded(format: OutputFormat) -> TaskContext {
-        let img: ImageBuffer<Rgba<u8>, Vec<u8>> =
-            ImageBuffer::from_pixel(4, 4, Rgba([10, 20, 30, 255]));
-        let dyn_img = DynamicImage::ImageRgba8(img);
-        TaskContext {
-            source: None,
-            decoded: Some(Arc::new(dyn_img)),
-            ops: vec![Operation::Resize {
-                width: Some(2),
-                height: Some(2),
-                fit: ResizeFit::Inside,
-            }],
-            format,
-            icc_profile: None,
-            icc_present: false,
-            exif_data: None,
-            auto_orient: true,
-            metadata_policy: MetadataPolicy::default_policy(),
-            firewall: FirewallConfig::disabled(),
-        }
-    }
-
-    fn make_task_with_decoded(format: OutputFormat) -> EncodeTask {
-        EncodeTask {
-            ctx: make_ctx_with_decoded(format),
-        }
-    }
-
-    #[test]
-    fn process_and_encode_outputs_image() {
-        let task = make_task_with_decoded(OutputFormat::Png);
-        let encoded = task
-            .process_and_encode(None)
-            .expect("encode should succeed");
-        assert!(!encoded.is_empty());
-    }
-
-    #[test]
-    fn decode_internal_errors_when_source_missing() {
-        let task = EncodeTask {
-            ctx: TaskContext {
-                source: None,
-                decoded: None,
-                ops: vec![],
-                format: OutputFormat::Png,
-                icc_profile: None,
-                icc_present: false,
-                exif_data: None,
-                auto_orient: true,
-                metadata_policy: MetadataPolicy::default_policy(),
-                firewall: FirewallConfig::disabled(),
-            },
-        };
-        let err = task.decode().unwrap_err();
-        assert!(matches!(err, LazyImageError::SourceConsumed));
-    }
-
-    /// Verify that base dimension/pixel limits are enforced at decode time
-    /// even when the firewall is disabled (i.e., sanitize() was NOT called).
-    /// This is the core fix for #492.
-    #[test]
-    fn base_limits_enforced_without_sanitize() {
-        // Create a PNG with dimensions exceeding MAX_PIXELS (100M)
-        // We can't easily create such a huge image, so instead verify via
-        // enforce_base_limits which is called in the decode path.
-        let cfg = FirewallConfig::disabled();
-        assert!(!cfg.enabled);
-
-        // 10001 x 10001 = 100_020_001 > 100M pixels
-        let err = cfg.enforce_base_limits(10001, 10001).unwrap_err();
-        assert!(matches!(
-            err,
-            crate::error::LazyImageError::PixelCountExceedsLimit { .. }
-        ));
-    }
-
-    /// Verify that base metadata scan runs at decode time without sanitize().
-    #[test]
-    fn base_metadata_scan_runs_without_sanitize() {
-        let cfg = FirewallConfig::disabled();
-        assert!(!cfg.enabled);
-
-        // Small metadata passes
-        let img: ImageBuffer<Rgba<u8>, Vec<u8>> =
-            ImageBuffer::from_pixel(4, 4, Rgba([10, 20, 30, 255]));
-        let mut bytes = Vec::new();
-        img.write_to(&mut std::io::Cursor::new(&mut bytes), ImageFormat::Png)
-            .unwrap();
-        assert!(cfg.scan_metadata_base(&bytes).is_ok());
-    }
-
-    #[test]
-    fn firewall_limits_are_enforced_on_decode() {
-        let png = sample_png_bytes();
-        let mut firewall = FirewallConfig::custom();
-        firewall.max_bytes = Some(1); // smaller than PNG size to force rejection
-
-        let task = EncodeTask {
-            ctx: TaskContext {
-                source: Some(Source::from_vec(png)),
-                decoded: None,
-                ops: vec![],
-                format: OutputFormat::Png,
-                icc_profile: None,
-                icc_present: false,
-                exif_data: None,
-                auto_orient: true,
-                metadata_policy: MetadataPolicy::default_policy(),
-                firewall,
-            },
-        };
-        let err = task.decode().unwrap_err();
-        assert!(matches!(err, LazyImageError::FirewallViolation { .. }));
-    }
-
-    #[test]
-    fn auto_orient_flag_disables_orientation_lookup() {
-        // When auto_orient is false, process_and_encode should NOT insert
-        // an AutoOrient operation even if the source has EXIF orientation.
-        // We test this indirectly: a PNG source has no EXIF, so orientation
-        // is None regardless. But we verify the flag is stored correctly.
-        let png = sample_png_bytes();
-        let task = EncodeTask {
-            ctx: TaskContext {
-                source: Some(Source::from_vec(png)),
-                decoded: None,
-                ops: vec![],
-                format: OutputFormat::Png,
-                icc_profile: None,
-                icc_present: false,
-                exif_data: None,
-                auto_orient: false,
-                metadata_policy: MetadataPolicy::default_policy(),
-                firewall: FirewallConfig::disabled(),
-            },
-        };
-        assert!(!task.ctx.auto_orient);
-        // process_and_encode should succeed (no orientation applied)
-        let result = task.process_and_encode(None);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn auto_orient_flag_default_is_true() {
-        // Verify the default task helper sets auto_orient to true
-        let task = make_task_with_decoded(OutputFormat::Png);
-        assert!(task.ctx.auto_orient);
-    }
-
-    #[test]
-    fn fast_mode_propagates_to_jpeg_encode() {
-        // Verify that OutputFormat::Jpeg with fast_mode=true produces valid JPEG
-        let task = make_task_with_decoded(OutputFormat::Jpeg {
-            quality: 80,
-            fast_mode: true,
-        });
-        let result = task
-            .process_and_encode(None)
-            .expect("encode should succeed");
-        // JPEG magic bytes
-        assert_eq!(&result[0..2], &[0xFF, 0xD8]);
-        assert_eq!(&result[result.len() - 2..], &[0xFF, 0xD9]);
-    }
-
-    #[test]
-    fn fast_mode_false_produces_valid_jpeg() {
-        let task = make_task_with_decoded(OutputFormat::Jpeg {
-            quality: 80,
-            fast_mode: false,
-        });
-        let result = task
-            .process_and_encode(None)
-            .expect("encode should succeed");
-        assert_eq!(&result[0..2], &[0xFF, 0xD8]);
-    }
-
-    #[test]
-    fn fast_mode_true_vs_false_both_valid() {
-        let fast_task = make_task_with_decoded(OutputFormat::Jpeg {
-            quality: 80,
-            fast_mode: true,
-        });
-        let normal_task = make_task_with_decoded(OutputFormat::Jpeg {
-            quality: 80,
-            fast_mode: false,
-        });
-        let fast_result = fast_task.process_and_encode(None).unwrap();
-        let normal_result = normal_task.process_and_encode(None).unwrap();
-        // Both should produce valid JPEGs
-        assert_eq!(&fast_result[0..2], &[0xFF, 0xD8]);
-        assert_eq!(&normal_result[0..2], &[0xFF, 0xD8]);
-        // Both should be non-empty
-        assert!(!fast_result.is_empty());
-        assert!(!normal_result.is_empty());
-    }
-
-    #[test]
-    fn target_bytes_search_finds_best_quality_under_budget() {
-        let img: ImageBuffer<Rgba<u8>, Vec<u8>> =
-            ImageBuffer::from_pixel(64, 64, Rgba([10, 20, 30, 255]));
-        let dyn_img = DynamicImage::ImageRgba8(img);
-
-        // Encode at max quality to get a reference size
-        let task_max = EncodeTask {
-            ctx: TaskContext {
-                source: None,
-                decoded: Some(Arc::new(dyn_img.clone())),
-                ops: vec![],
-                format: OutputFormat::Jpeg {
-                    quality: 100,
-                    fast_mode: false,
-                },
-                icc_profile: None,
-                icc_present: false,
-                exif_data: None,
-                auto_orient: false,
-                metadata_policy: MetadataPolicy::strip_all(),
-                firewall: FirewallConfig::disabled(),
-            },
-        };
-        let max_size = task_max.process_and_encode(None).unwrap().len();
-
-        // Set a budget larger than any output (should pick max quality)
-        let search_task = EncodeTargetBytesTask {
-            ctx: TaskContext {
-                source: None,
-                decoded: Some(Arc::new(dyn_img.clone())),
-                ops: vec![],
-                format: OutputFormat::Jpeg {
-                    quality: 100,
-                    fast_mode: false,
-                },
-                icc_profile: None,
-                icc_present: false,
-                exif_data: None,
-                auto_orient: false,
-                metadata_policy: MetadataPolicy::strip_all(),
-                firewall: FirewallConfig::disabled(),
-            },
-            target_bytes: (max_size + 10000) as u32,
-            min_quality: 1,
-            max_quality: 100,
-            quality_floor_policy_strict: false,
-        };
-        let result = search_task.search().expect("search should succeed");
-        assert!(result.budget_met);
-        assert_eq!(result.quality, 100);
-
-        // Set a tight budget (should pick a lower quality)
-        let tight_task = EncodeTargetBytesTask {
-            ctx: TaskContext {
-                source: None,
-                decoded: Some(Arc::new(dyn_img)),
-                ops: vec![],
-                format: OutputFormat::Jpeg {
-                    quality: 100,
-                    fast_mode: false,
-                },
-                icc_profile: None,
-                icc_present: false,
-                exif_data: None,
-                auto_orient: false,
-                metadata_policy: MetadataPolicy::strip_all(),
-                firewall: FirewallConfig::disabled(),
-            },
-            target_bytes: 500,
-            min_quality: 1,
-            max_quality: 100,
-            quality_floor_policy_strict: false,
-        };
-        let tight_result = tight_task.search().expect("search should succeed");
-        // Quality should be lower than max since budget is tight
-        assert!(tight_result.quality <= 100);
-        // The result should be the best effort (either under budget or closest over)
-        assert!(!tight_result.data.is_empty());
-    }
-
-    #[test]
-    fn target_bytes_strict_policy_fails_when_budget_unmet() {
-        let img: ImageBuffer<Rgba<u8>, Vec<u8>> =
-            ImageBuffer::from_pixel(64, 64, Rgba([10, 20, 30, 255]));
-        let dyn_img = DynamicImage::ImageRgba8(img);
-
-        let task = EncodeTargetBytesTask {
-            ctx: TaskContext {
-                source: None,
-                decoded: Some(Arc::new(dyn_img)),
-                ops: vec![],
-                format: OutputFormat::Jpeg {
-                    quality: 100,
-                    fast_mode: false,
-                },
-                icc_profile: None,
-                icc_present: false,
-                exif_data: None,
-                auto_orient: false,
-                metadata_policy: MetadataPolicy::strip_all(),
-                firewall: FirewallConfig::disabled(),
-            },
-            // Impossibly small budget
-            target_bytes: 1,
-            min_quality: 80,
-            max_quality: 100,
-            quality_floor_policy_strict: true,
-        };
-        let err = task.search().unwrap_err();
-        let msg = err.to_string();
-        assert!(
-            msg.contains("Unable to meet targetBytes"),
-            "expected strict policy error, got: {msg}"
-        );
-    }
-
-    /// Regression test for issue #558: the binary search must decode + apply
-    /// ops once, not per quality iteration. We assert it indirectly by giving
-    /// the prepare phase a pre-decoded Arc and checking that `prepare_for_encode`
-    /// reuses the Arc by reference (no pixel clone) when ops is empty.
-    #[test]
-    fn prepare_for_encode_reuses_decoded_arc_when_ops_empty() {
-        let img: ImageBuffer<Rgba<u8>, Vec<u8>> =
-            ImageBuffer::from_pixel(32, 32, Rgba([10, 20, 30, 255]));
-        let dyn_img = DynamicImage::ImageRgba8(img);
-        let original = Arc::new(dyn_img);
-
-        let prepared = prepare_for_encode(
-            None,
-            Some(&original),
-            &[],
-            false,
-            false,
-            &FirewallConfig::disabled(),
-            &OutputFormat::Jpeg {
-                quality: 80,
-                fast_mode: false,
-            },
-        )
-        .expect("prepare should succeed");
-
-        // Cheap path: Arc clone, not a deep image clone.
-        assert!(
-            Arc::ptr_eq(&original, &prepared.processed),
-            "prepare_for_encode must reuse the decoded Arc when no ops are applied"
-        );
-    }
-
-    /// Regression test for issue #558: ops *are* applied during prepare, and
-    /// the resulting Arc points to a freshly-allocated transformed image
-    /// (not the original input).
-    #[test]
-    fn prepare_for_encode_applies_ops_once_when_ops_nonempty() {
-        let img: ImageBuffer<Rgba<u8>, Vec<u8>> =
-            ImageBuffer::from_pixel(64, 64, Rgba([10, 20, 30, 255]));
-        let dyn_img = DynamicImage::ImageRgba8(img);
-        let original = Arc::new(dyn_img);
-
-        let prepared = prepare_for_encode(
-            None,
-            Some(&original),
-            &[Operation::Resize {
-                width: Some(16),
-                height: Some(16),
-                fit: crate::ops::ResizeFit::Inside,
-            }],
-            false,
-            false,
-            &FirewallConfig::disabled(),
-            &OutputFormat::Jpeg {
-                quality: 80,
-                fast_mode: false,
-            },
-        )
-        .expect("prepare should succeed");
-
-        assert!(
-            !Arc::ptr_eq(&original, &prepared.processed),
-            "prepare_for_encode must produce a new image when ops transform pixels"
-        );
-        assert_eq!(prepared.processed.width(), 16);
-        assert_eq!(prepared.processed.height(), 16);
-    }
-
-    /// Regression test for issue #558: re-running encode_prepared multiple
-    /// times against the same PreparedEncode must produce consistent results,
-    /// proving the cached image is reused safely.
-    #[test]
-    fn encode_prepared_can_run_multiple_times_against_same_cache() {
-        let img: ImageBuffer<Rgba<u8>, Vec<u8>> =
-            ImageBuffer::from_pixel(64, 64, Rgba([200, 100, 50, 255]));
-        let dyn_img = DynamicImage::ImageRgba8(img);
-
-        let prepared = prepare_for_encode(
-            None,
-            Some(&Arc::new(dyn_img)),
-            &[],
-            false,
-            false,
-            &FirewallConfig::disabled(),
-            &OutputFormat::Jpeg {
-                quality: 80,
-                fast_mode: false,
-            },
-        )
-        .expect("prepare should succeed");
-
-        // Encode at three different qualities — like the binary search would.
-        let mut last_size: Option<usize> = None;
-        let policy = MetadataPolicy::strip_all();
-        let firewall = FirewallConfig::disabled();
-        for quality in [50u8, 80, 100] {
-            let format = OutputFormat::Jpeg {
-                quality,
-                fast_mode: false,
-            };
-            let ctx = EncodeContext {
-                source: None,
-                decoded: None,
-                ops: &[],
-                format: &format,
-                icc_profile: None,
-                icc_present: false,
-                exif_data: None,
-                auto_orient: false,
-                policy: &policy,
-                firewall: &firewall,
-            };
-            let bytes = encode_prepared(&prepared, &ctx, None).expect("encode should succeed");
-            assert_eq!(
-                &bytes[0..2],
-                &[0xFF, 0xD8],
-                "encode_prepared must produce valid JPEG at quality {quality}"
-            );
-            // Higher quality should generally yield larger files.
-            if let Some(prev) = last_size {
-                assert!(
-                    bytes.len() >= prev,
-                    "higher quality should not produce smaller output: prev={prev}, current={}",
-                    bytes.len()
-                );
-            }
-            last_size = Some(bytes.len());
-        }
-    }
-
-    /// Regression test for issue #558: the metrics returned by the byte-budget
-    /// search must include sensible decode_ms / ops_ms / encode_ms fields
-    /// (the prepare phase populates the first two; encode populates the third).
-    #[test]
-    fn target_bytes_search_metrics_include_decode_ops_and_encode_timings() {
-        let img: ImageBuffer<Rgba<u8>, Vec<u8>> =
-            ImageBuffer::from_pixel(64, 64, Rgba([10, 20, 30, 255]));
-        let dyn_img = DynamicImage::ImageRgba8(img);
-
-        let task = EncodeTargetBytesTask {
-            ctx: TaskContext {
-                source: None,
-                decoded: Some(Arc::new(dyn_img)),
-                ops: vec![Operation::Resize {
-                    width: Some(32),
-                    height: Some(32),
-                    fit: crate::ops::ResizeFit::Inside,
-                }],
-                format: OutputFormat::Jpeg {
-                    quality: 100,
-                    fast_mode: false,
-                },
-                icc_profile: None,
-                icc_present: false,
-                exif_data: None,
-                auto_orient: false,
-                metadata_policy: MetadataPolicy::strip_all(),
-                firewall: FirewallConfig::disabled(),
-            },
-            target_bytes: 10_000_000,
-            min_quality: 1,
-            max_quality: 100,
-            quality_floor_policy_strict: false,
-        };
-
-        let result = task.search().expect("search should succeed");
-        let metrics = result.metrics;
-
-        // All stage timings must be non-negative (Instant::elapsed cannot be negative
-        // but the f64 cast can technically produce 0.0; assert >=).
-        assert!(metrics.decode_ms >= 0.0);
-        assert!(metrics.ops_ms >= 0.0);
-        assert!(metrics.encode_ms >= 0.0);
-        assert!(metrics.total_ms >= metrics.encode_ms);
-        // bytes_in/out should reflect the actual output (no input bytes available).
-        assert_eq!(metrics.bytes_in, 0);
-        assert_eq!(metrics.bytes_out as usize, result.data.len());
-    }
-}
