+++ src/engine/tasks/batch.rs
@@ -1,0 +1,490 @@
+// src/engine/tasks/batch.rs
+//
+// Batch processing tasks: process many input files in parallel with rayon.
+//
+// `BatchTask` and `BatchWithMetricsTask` are kept as hand-written `Task` impls
+// (not generated by `define_encode_task!`) because their concurrency model
+// differs structurally — they own their own input list, manage rayon workers,
+// and collect per-file results rather than running a single encode pipeline.
+
+use super::super::api::MetadataPolicy;
+use super::super::firewall::FirewallConfig;
+use super::atomic::write_encoded_output;
+#[cfg(feature = "napi")]
+use super::context::BatchResult;
+use super::processing::{process_and_encode_from_parts, EncodeContext};
+use crate::engine::io::{extract_exif_raw, extract_icc_profile};
+#[cfg(feature = "napi")]
+use crate::engine::pool;
+use crate::error::LazyImageError;
+use crate::ops::{Operation, OutputFormat};
+#[cfg(feature = "napi")]
+use napi::bindgen_prelude::*;
+#[cfg(feature = "napi")]
+use napi::{Env, Task};
+#[cfg(feature = "napi")]
+use rayon::prelude::*;
+
+struct BatchFileSuccess {
+    output_path: String,
+    metrics: Option<crate::ProcessingMetrics>,
+}
+
+/// Borrowed parameter bundle for [`process_batch_file`].
+///
+/// Holds the per-batch configuration (output directory, ops, format, policy,
+/// firewall, etc.) that is identical for every file in a batch. Built once
+/// outside the rayon loop and shared (by `Copy`) across every worker closure.
+#[derive(Clone, Copy)]
+struct BatchFileContext<'a> {
+    output_dir: &'a str,
+    ops: &'a [Operation],
+    format: &'a OutputFormat,
+    auto_orient: bool,
+    policy: &'a MetadataPolicy,
+    firewall: &'a FirewallConfig,
+    collect_metrics: bool,
+}
+
+fn process_batch_file(
+    input_path: &str,
+    ctx: &BatchFileContext<'_>,
+) -> std::result::Result<BatchFileSuccess, LazyImageError> {
+    let keep_icc = ctx.policy.effective_icc();
+    let keep_exif = ctx.policy.effective_exif();
+    use std::path::Path;
+    use std::sync::Arc;
+
+    let input_len = std::fs::metadata(input_path)
+        .map_err(|e| {
+            if e.kind() == std::io::ErrorKind::NotFound {
+                LazyImageError::file_not_found(input_path.to_string())
+            } else {
+                LazyImageError::file_read_failed(input_path.to_string(), e)
+            }
+        })?
+        .len() as usize;
+    ctx.firewall.enforce_source_len(input_len)?;
+
+    let source = crate::engine::io::load_file_safe(Path::new(input_path))?;
+    let data = source.as_bytes().expect("source always has bytes");
+
+    ctx.firewall.scan_metadata(data)?;
+
+    let extracted_icc = extract_icc_profile(data)?;
+    let icc_present = extracted_icc.is_some();
+    let icc_profile = if keep_icc {
+        extracted_icc.map(Arc::new)
+    } else {
+        None
+    };
+    let exif_data = if keep_exif {
+        extract_exif_raw(data).map(Arc::new)
+    } else {
+        None
+    };
+
+    let mut metrics = ctx.collect_metrics.then(crate::ProcessingMetrics::default);
+    let encode_ctx = EncodeContext {
+        source: Some(&source),
+        decoded: None,
+        ops: ctx.ops,
+        format: ctx.format,
+        icc_profile: icc_profile.as_ref(),
+        icc_present,
+        exif_data: exif_data.as_ref(),
+        auto_orient: ctx.auto_orient,
+        policy: ctx.policy,
+        firewall: ctx.firewall,
+    };
+    let encoded = process_and_encode_from_parts(&encode_ctx, metrics.as_mut())?;
+
+    let filename = Path::new(input_path)
+        .file_name()
+        .ok_or_else(|| LazyImageError::internal_panic("invalid filename"))?;
+    let extension = match ctx.format {
+        OutputFormat::Jpeg { .. } => "jpg",
+        OutputFormat::Png => "png",
+        OutputFormat::WebP { .. } => "webp",
+        OutputFormat::Avif { .. } => "avif",
+    };
+    let output_filename = Path::new(filename).with_extension(extension);
+    let output_path = Path::new(ctx.output_dir).join(output_filename);
+    write_encoded_output(output_path.to_string_lossy().as_ref(), &encoded)?;
+
+    Ok(BatchFileSuccess {
+        output_path: output_path.to_string_lossy().to_string(),
+        metrics,
+    })
+}
+
+pub struct BatchTask {
+    pub inputs: Vec<String>,
+    pub output_dir: String,
+    pub ops: Vec<Operation>,
+    pub format: OutputFormat,
+    pub concurrency: u32,
+    pub auto_orient: bool,
+    /// Single source of truth for metadata preservation decisions
+    pub metadata_policy: MetadataPolicy,
+    pub firewall: FirewallConfig,
+    /// Last error that occurred during compute (for use in reject)
+    #[cfg(feature = "napi")]
+    pub(crate) last_error: Option<LazyImageError>,
+}
+
+pub struct BatchWithMetricsTask {
+    pub inputs: Vec<String>,
+    pub output_dir: String,
+    pub ops: Vec<Operation>,
+    pub format: OutputFormat,
+    pub concurrency: u32,
+    pub auto_orient: bool,
+    /// Single source of truth for metadata preservation decisions
+    pub metadata_policy: MetadataPolicy,
+    pub firewall: FirewallConfig,
+    #[cfg(feature = "napi")]
+    pub(crate) last_error: Option<LazyImageError>,
+}
+
+// Compile-time `Send` assertions for the manually-implemented batch tasks.
+// Matches the assertion the `define_encode_task!` macro emits for the other
+// task types — see the comment near the macro definition.
+const _: fn() = || {
+    fn assert_send<T: Send>() {}
+    assert_send::<BatchTask>();
+    assert_send::<BatchWithMetricsTask>();
+};
+
+#[cfg(feature = "napi")]
+#[napi]
+impl Task for BatchTask {
+    type Output = Vec<BatchResult>;
+    type JsValue = Vec<BatchResult>;
+
+    fn compute(&mut self) -> Result<Self::Output> {
+        use std::fs;
+        use std::path::Path;
+
+        if !Path::new(&self.output_dir).exists() {
+            fs::create_dir_all(&self.output_dir).map_err(|e| {
+                let lazy_err = LazyImageError::file_write_failed(self.output_dir.clone(), e);
+                self.last_error = Some(lazy_err.clone());
+                napi::Error::from(lazy_err)
+            })?;
+        }
+
+        // concurrency = 0 means "use default" (auto-detected via CPU and memory)
+        // concurrency = 1..MAX_CONCURRENCY means "use specified number of concurrent operations"
+        if self.concurrency > pool::MAX_CONCURRENCY as u32 {
+            let lazy_err = LazyImageError::invalid_argument(
+                "concurrency",
+                self.concurrency.to_string(),
+                format!("must be 0 or 1-{}", pool::MAX_CONCURRENCY),
+            );
+            self.last_error = Some(lazy_err.clone());
+            return Err(napi::Error::from(lazy_err));
+        }
+
+        let effective_concurrency = pool::resolve_effective_concurrency(self.concurrency);
+        let effective_concurrency_u32 = u32::try_from(effective_concurrency).ok();
+        let auto_concurrency = Some(self.concurrency == 0);
+
+        // Helper closure to process a single image
+        let firewall = self.firewall.clone();
+        let batch_ctx = BatchFileContext {
+            output_dir: &self.output_dir,
+            ops: &self.ops,
+            format: &self.format,
+            auto_orient: self.auto_orient,
+            policy: &self.metadata_policy,
+            firewall: &firewall,
+            collect_metrics: false,
+        };
+        let process_one = |input_path: &String| -> BatchResult {
+            let result = process_batch_file(input_path, &batch_ctx);
+
+            match result {
+                Ok(output) => BatchResult {
+                    source: input_path.clone(),
+                    success: true,
+                    error: None,
+                    output_path: Some(output.output_path),
+                    error_code: None,
+                    error_category: None,
+                    effective_concurrency: effective_concurrency_u32,
+                    auto_concurrency,
+                },
+                Err(err) => {
+                    let error_code = err.code();
+                    let error_msg = format!("[{}] {}: {}", error_code.as_str(), input_path, err);
+                    let category = error_code.category();
+                    BatchResult {
+                        source: input_path.clone(),
+                        success: false,
+                        error: Some(error_msg),
+                        output_path: None,
+                        error_code: Some(error_code.as_str().to_string()),
+                        error_category: Some(category),
+                        effective_concurrency: effective_concurrency_u32,
+                        auto_concurrency,
+                    }
+                }
+            }
+        };
+
+        // Use the shared rayon pool to avoid per-call pool construction overhead.
+        // Memory backpressure is still handled by WeightedSemaphore.
+        //
+        // When effective_concurrency < inputs.len(), run exactly that many worker
+        // tasks on the global pool and pull work dynamically via an atomic index.
+        // This preserves the concurrency cap while avoiding chunk-induced tail latency.
+        let pool = match pool::get_pool() {
+            Ok(pool) => pool,
+            Err(err) => {
+                self.last_error = Some(err.clone());
+                return Err(napi::Error::from(err));
+            }
+        };
+
+        let results: Vec<BatchResult> = if effective_concurrency >= self.inputs.len() {
+            pool.install(|| self.inputs.par_iter().map(process_one).collect())
+        } else {
+            use std::sync::atomic::{AtomicUsize, Ordering};
+            use std::sync::{Arc, Mutex};
+
+            let next_index = Arc::new(AtomicUsize::new(0));
+            let indexed_results = Arc::new(Mutex::new(Vec::<(usize, BatchResult)>::with_capacity(
+                self.inputs.len(),
+            )));
+
+            pool.install(|| {
+                rayon::scope(|scope| {
+                    for _ in 0..effective_concurrency {
+                        let next_index = Arc::clone(&next_index);
+                        let indexed_results = Arc::clone(&indexed_results);
+                        let inputs = &self.inputs;
+                        scope.spawn(move |_| loop {
+                            let idx = next_index.fetch_add(1, Ordering::Relaxed);
+                            if idx >= inputs.len() {
+                                break;
+                            }
+
+                            let result = process_one(&inputs[idx]);
+                            let mut guard = indexed_results
+                                .lock()
+                                .unwrap_or_else(|poisoned| poisoned.into_inner());
+                            guard.push((idx, result));
+                        });
+                    }
+                });
+            });
+
+            let mut indexed_results = {
+                let mut guard = indexed_results
+                    .lock()
+                    .unwrap_or_else(|poisoned| poisoned.into_inner());
+                std::mem::take(&mut *guard)
+            };
+            indexed_results.sort_unstable_by_key(|(idx, _)| *idx);
+            indexed_results
+                .into_iter()
+                .map(|(_, result)| result)
+                .collect()
+        };
+
+        Ok(results)
+    }
+
+    fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
+        Ok(output)
+    }
+
+    fn reject(&mut self, env: Env, err: napi::Error) -> Result<Self::JsValue> {
+        // Use stored error if available, otherwise try to extract from napi::Error
+        let lazy_err = self.last_error.take().unwrap_or_else(|| {
+            // Fallback: create a generic error from the napi::Error message
+            // This should not happen if all error paths properly preserve LazyImageError
+            LazyImageError::generic(err.to_string())
+        });
+        let napi_err = crate::error::napi_error_with_code(&env, lazy_err)?;
+        Err(napi_err)
+    }
+}
+
+#[cfg(feature = "napi")]
+#[napi]
+impl Task for BatchWithMetricsTask {
+    type Output = crate::BatchOutputWithMetrics;
+    type JsValue = crate::BatchOutputWithMetrics;
+
+    fn compute(&mut self) -> Result<Self::Output> {
+        use std::fs;
+        use std::path::Path;
+
+        if !Path::new(&self.output_dir).exists() {
+            fs::create_dir_all(&self.output_dir).map_err(|e| {
+                let lazy_err = LazyImageError::file_write_failed(self.output_dir.clone(), e);
+                self.last_error = Some(lazy_err.clone());
+                napi::Error::from(lazy_err)
+            })?;
+        }
+
+        if self.concurrency > pool::MAX_CONCURRENCY as u32 {
+            let lazy_err = LazyImageError::invalid_argument(
+                "concurrency",
+                self.concurrency.to_string(),
+                format!("must be 0 or 1-{}", pool::MAX_CONCURRENCY),
+            );
+            self.last_error = Some(lazy_err.clone());
+            return Err(napi::Error::from(lazy_err));
+        }
+
+        let effective_concurrency = pool::resolve_effective_concurrency(self.concurrency);
+        let effective_concurrency_u32 = u32::try_from(effective_concurrency).unwrap_or(u32::MAX);
+        let auto_concurrency = self.concurrency == 0;
+
+        let firewall = self.firewall.clone();
+        let batch_ctx = BatchFileContext {
+            output_dir: &self.output_dir,
+            ops: &self.ops,
+            format: &self.format,
+            auto_orient: self.auto_orient,
+            policy: &self.metadata_policy,
+            firewall: &firewall,
+            collect_metrics: true,
+        };
+
+        let process_one = |input_path: &String| -> crate::BatchResultWithMetrics {
+            match process_batch_file(input_path, &batch_ctx) {
+                Ok(output) => crate::BatchResultWithMetrics {
+                    source: input_path.clone(),
+                    success: true,
+                    error: None,
+                    output_path: Some(output.output_path),
+                    error_code: None,
+                    error_category: None,
+                    effective_concurrency: Some(effective_concurrency_u32),
+                    auto_concurrency: Some(auto_concurrency),
+                    metrics: output.metrics,
+                },
+                Err(err) => {
+                    let error_code = err.code();
+                    let error_msg = format!("[{}] {}: {}", error_code.as_str(), input_path, err);
+                    let category = error_code.category();
+                    crate::BatchResultWithMetrics {
+                        source: input_path.clone(),
+                        success: false,
+                        error: Some(error_msg),
+                        output_path: None,
+                        error_code: Some(error_code.as_str().to_string()),
+                        error_category: Some(category),
+                        effective_concurrency: Some(effective_concurrency_u32),
+                        auto_concurrency: Some(auto_concurrency),
+                        metrics: None,
+                    }
+                }
+            }
+        };
+
+        let pool = match pool::get_pool() {
+            Ok(pool) => pool,
+            Err(err) => {
+                self.last_error = Some(err.clone());
+                return Err(napi::Error::from(err));
+            }
+        };
+
+        let items: Vec<crate::BatchResultWithMetrics> =
+            if effective_concurrency >= self.inputs.len() {
+                pool.install(|| self.inputs.par_iter().map(process_one).collect())
+            } else {
+                use std::sync::atomic::{AtomicUsize, Ordering};
+                use std::sync::{Arc, Mutex};
+
+                let inputs = Arc::new(&self.inputs);
+                let out = Arc::new(Mutex::new(Vec::with_capacity(self.inputs.len())));
+                let next = AtomicUsize::new(0);
+
+                pool.install(|| {
+                    rayon::scope(|s| {
+                        for _ in 0..effective_concurrency {
+                            let inputs = Arc::clone(&inputs);
+                            let out = Arc::clone(&out);
+                            let next_ref = &next;
+                            s.spawn(move |_| loop {
+                                let idx = next_ref.fetch_add(1, Ordering::Relaxed);
+                                if idx >= inputs.len() {
+                                    break;
+                                }
+                                let item = process_one(&inputs[idx]);
+                                out.lock().unwrap().push((idx, item));
+                            });
+                        }
+                    });
+                });
+
+                let mut pairs = Arc::try_unwrap(out).unwrap().into_inner().unwrap();
+                pairs.sort_by_key(|(idx, _)| *idx);
+                pairs.into_iter().map(|(_, item)| item).collect()
+            };
+
+        let mut successful_items = 0u32;
+        let mut failed_items = 0u32;
+        let mut total_bytes_in = 0f64;
+        let mut total_bytes_out = 0f64;
+        let mut total_decode_ms = 0f64;
+        let mut total_ops_ms = 0f64;
+        let mut total_encode_ms = 0f64;
+        let mut total_cpu_time = 0f64;
+        let mut total_wall_ms = 0f64;
+
+        for item in &items {
+            if item.success {
+                successful_items += 1;
+            } else {
+                failed_items += 1;
+            }
+            if let Some(metrics) = item.metrics.as_ref() {
+                total_bytes_in += metrics.bytes_in as f64;
+                total_bytes_out += metrics.bytes_out as f64;
+                total_decode_ms += metrics.decode_ms;
+                total_ops_ms += metrics.ops_ms;
+                total_encode_ms += metrics.encode_ms;
+                total_cpu_time += metrics.cpu_time;
+                total_wall_ms += metrics.total_ms;
+            }
+        }
+
+        Ok(crate::BatchOutputWithMetrics {
+            items,
+            summary: crate::BatchMetricsSummary {
+                total_items: self.inputs.len() as u32,
+                successful_items,
+                failed_items,
+                effective_concurrency: effective_concurrency_u32,
+                auto_concurrency,
+                total_bytes_in,
+                total_bytes_out,
+                total_decode_ms,
+                total_ops_ms,
+                total_encode_ms,
+                total_cpu_time,
+                total_wall_ms,
+            },
+        })
+    }
+
+    fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
+        Ok(output)
+    }
+
+    fn reject(&mut self, env: Env, err: napi::Error) -> Result<Self::JsValue> {
+        let lazy_err = self
+            .last_error
+            .take()
+            .unwrap_or_else(|| LazyImageError::generic(err.to_string()));
+        let napi_err = crate::error::napi_error_with_code(&env, lazy_err)?;
+        Err(napi_err)
+    }
+}
